ColorThis/requirements.txt
Keras==2.2.5
Keras-Preprocessing==1.1.0
matplotlib-venn==0.11.5
numpy==1.17.4
Pillow==6.2.1
scikit-image==0.15.0
tensorflow-gpu==1.15.0
